www/app/requirements.txt
uvicorn
fastapi[standard]
opentelemetry-sdk